+++ index.html
@@ -11,8 +11,7 @@
 </head>
 <body>
 
-<script src="//cdn.jsdelivr.net/npm/phaser@3.55.2/dist/phaser.js"></script>
-<script src="src/Game.js"></script>
+<script type="module" src="/src/index.ts"></script>
 
 </body>
 </html>
